LayoutTests/media/modern-media-controls/button/button.html

<link rel="stylesheet" href="../../../../Source/WebCore/Modules/modern-media-controls/controls/button.css" type="text/css" media="screen">
<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>Button</code> class.");

debug("Default values");
const button = new Button;
shouldBeTrue("button.enabled");
shouldBeUndefined("button.uiDelegate");
shouldBeEqualToString("button.element.localName", "button");

let clickedButton;

debug("");
debug("Responds to click events");
button.uiDelegate = {

    buttonWasPressed: function(aButton)
    {
        clickedButton = aButton;
        shouldBeTrue("clickedButton === button");
    }

}
button.element.dispatchEvent(new MouseEvent("click"));

debug("");
debug("Doesn't respond to click events when disabled");
button.enabled = false;
button.element.dispatchEvent(new MouseEvent("click"));

debug("");
debug("Has all the appropriate styles")
document.body.appendChild(button.element);
const style = window.getComputedStyle(button.element);
shouldBeEqualToString("style.position", "absolute");
shouldBeEqualToString("style.borderWidth", "0px");
shouldBeEqualToString("style.webkitAppearance", "none");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
